move dhcp code to common file
